Backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg[binary]==3.3.3
python-dotenv==1.0.0
web3==7.10.0
